Initial commit for #120 creating a simple CSV exporter
This contains already all basic functionality but is lacking a test case and some extra tests.
The `ben-csv` utility was added. It takes an optional argument `-o` to output to csv file (otherwise stdout) and a campaign directory. It then uses `hpcbench.campaign` utilities to iterate over all benchmarks in a campaign and return all data in csv format.
Currently this will return all possible columns. So if multiple benhcmarks are in one campaign, the table might get rather wide.