Added a switch for environment variable expansion in campaign files
Post-processing tools such as `ben-elastic` or `ben-csv` cannot rely on environment variable expansion to work properly as they can be executed on a campaign in a completely different environment.
Ideally, the environment at execution should be captured and reapplied when the campaign file is parsed.
This commit proposed a simpler solution, which should be acceptable for most situations: disable environment variable expansion when we are simply postprocessing an hpcbench report.
This should close #116 since it fixes the associated issue.
expanding variables inside the written campaign file brings other issues, which are harder to address in some of the postprocessing tools.