Can now extract outputs of every SLURM processes
Standard output and error of every processes spawned by SLURM
are redirected to dedicated files. It allows hpcbench to extract
metrics from everyone of those.

Therefore, structre of the `metrics` key in both YAML and JSON outputs
has changed.
Before it was a dictionary of metrics name -> value. Now it provides
metrics for every outputs of the command, for instance:

```yaml
metrics:
  -
    context:
      node: r1n1
      rank: 1
      executor: slurm
    measurement:
      latency: 1.1
  -
    context:
      node: r1n1
      rank: 2
      executor: slurm
    measurement:
      latency: 3
```

When commands are spawned with local executor, results look like:

```yaml
metrics:
  -
    context:
      executor: local
    measurement:
      latency: 1.1
```
Also:
* Remove plotting capabilities because unused
* Remove `metrics` key in benchmark YAML configuration because should
  have not been done in the first place.