Don't expand the campaign when not in data generation mode
Neither environment variables nor network section should be expanded when we
are traversing the campaign for exporting, reporting, etc. These actions are
typically done at a later point in time and not necessarily on the same
system and would fail when we try to expand the original environment.
NB: we should consider storing the campaign file in expanded form in the
report directory, cf. #220