ExecutionContext: new benchmark attribute
`ExecutionContext` now has a `benchmark` string  attribute providing the
benchmark name specified in YAML to every `hpcbench.api.Benchmark` instances.

This change is minimalist in `hpcbench/` directory but to test it,
an assert is added to `FakeBenchmark.execution_matrix` to verify
the value of this new attribute.
Because `FakeBenchmark` is heavily used in `tests/` and the benchmark
name didn't matter before in YAMLs, it implies a lot of changes.